Revert "Save tour info"
This reverts commit 954ef4f1955c899aca7415fb4e47340bf5df333d.
