--- react/components/Tour/data.js
@@ -1,1 +1,0 @@
-tour
