+++ react/components/drawer.js
@@ -82,7 +82,6 @@ export default function Drawer({ children, element_id, classes, source_type, set
                                 filterQuery={state.filterQuery}
                                 filterStormName={state.filterStormName}
                                 showFilterSelected={state.showFilterSelected}
-                                isTour={state.isTour}
                                 
                         />
                         ) :
