+++ react/components/map.js
@@ -23,16 +23,13 @@ import { useMediaQuery, Box, useTheme } from "@mui/material";
 import TourWrapper from "@/components/Tour/UseTour";
 import { getTourSteps } from "@/components/Tour/tourSteps";
 import { useTour } from "@reactour/tour";
-import { useRouter } from 'next/router';
-import { TourProvider } from "@reactour/tour";
-
 
 
 const defaultPosition = [46.9736, -54.69528]; // Mouth of Placentia Bay
 const defaultZoom = 4
 
 
-export default function Map({ children, station_data, source_type,  setStationPoints}) {
+export default function Map({ children, station_data, source_type,  setStationPoints, isTourReady}) {
 
   const clearShapesRef = useRef(null);
 
@@ -41,36 +38,13 @@ export default function Map({ children, station_data, source_type,  setStationPo
   const theme = useTheme();
   const { setIsOpen, setCurrentStep } = useTour();
   const [showModal, setShowModal] = useState(true);
-  
-  const router = useRouter();
-  const routerReady = router?.isReady;
-
-
-
-  const mode = router?.query?.storms;
-    const hasStorms = Boolean(router?.query?.name);
-    const isActive = mode === "active";
-    const isHistorical = mode === "historical";
-    const hasStations =
-  state.station_data && Object.keys(state.station_data).length > 0;
-
-    const steps = useMemo(() => {
-
-      return getTourSteps({
-        hasStorms,
-        hasStations,
-        isActive,
-        isHistorical,
-      });
-    }, [hasStorms, hasStations, isActive, isHistorical]);
  
 
   
-  const startTour = async () => {
-  setShowModal(false);
-  await router.push(
-    "?storms=historical&name=FIONA&season=2022&sid=2022257N16312"
-    );
+  const startTour = () => {
+    setShowModal(false);
+    setCurrentStep(0);
+    setIsOpen(true);
     
   };
 
@@ -78,18 +52,6 @@ export default function Map({ children, station_data, source_type,  setStationPo
     setShowModal(false);
     
   };
-
-  useEffect(() => {
-    if (!router.isReady) return;
-
-    const mode = router.query.storms;
-    const hasStorms = Boolean(router.query.name);
-
-    if (mode && hasStorms) {
-      setCurrentStep(0);
-      setIsOpen(true);
-    }
-  }, [router.isReady, router.query]);
 
 
 
@@ -103,245 +65,237 @@ export default function Map({ children, station_data, source_type,  setStationPo
   
 
   return (
-    <TourProvider steps={steps}
-        styles={{
-          popover: base => ({
-            ...base,
-            zIndex: 10000000
-          })
-        }} >
     
-      <div className="map_container">
-        <div className='inner_container'>
-          
-        {showModal && (
-            <div className="tour-modal-overlay">
-              <div className="tour-modal">
-                <h2>Welcome</h2>
-                <p>
-                  Want a quick tour of how to explore storms and use the map?
-                </p>
+    <div className="map_container">
+      <div className='inner_container'>
+         
+       {showModal && (
+          <div className="tour-modal-overlay">
+            <div className="tour-modal">
+              <h2>Welcome 👋</h2>
+              <p>
+                Want a quick tour of how to explore storms and use the map?
+              </p>
 
-                <div className="tour-actions">
-                  <button
-                    onClick={startTour}
-                    className="primary"
-                    
-                  >
-                    Take a Tour
-                  </button>
+              <div className="tour-actions">
+                <button
+                  onClick={startTour}
+                  className="primary"
+                  disabled={!isTourReady}
+                >
+                  Take a Tour
+                </button>
 
-                  <button onClick={skipTour} className="secondary">
-                    No, thanks
-                  </button>
-                </div>
+                <button onClick={skipTour} className="secondary">
+                  No, thanks
+                </button>
               </div>
             </div>
-          )}
+          </div>
+        )}
 
-        
-          <IconButton
-            className="info-guide"
-            sx={{ display: 'flex'
-              }}
-            onClick={() => {
-              dispatch({ type: "SET_INFO_GUIDE", payload: true});
+      
+        <IconButton
+          className="info-guide"
+          sx={{ display: 'flex'
             }}
-            ><InfoIcon />
-          </IconButton>
+          onClick={() => {
+            dispatch({ type: "SET_INFO_GUIDE", payload: true});
+          }}
+          ><InfoIcon />
+        </IconButton>
 
+      
+      
         
-        
-          
-          { source_type === "historical" &&
-            <RenderFilter
-            clearShapesRef={clearShapesRef} // Pass the ref to 
+        { source_type === "historical" &&
+          <RenderFilter
+          clearShapesRef={clearShapesRef} // Pass the ref to 
+          state={state}
+          dispatch={dispatch}
+          setStationPoints={setStationPoints}
+          />
+        }
+        {
+          <RenderDashboards
+            source_type={source_type}
+            time = {new Date()}
             state={state}
             dispatch={dispatch}
-            setStationPoints={setStationPoints}
+            
             />
-          }
-          {
-            <RenderDashboards
-              source_type={source_type}
-              time = {new Date()}
-              state={state}
-              dispatch={dispatch}
-              
-              />
-          }
-          {//state.isDrawerOpen && <div className="drawer-touch-blocker"/>
-          }
+        }
+        {//state.isDrawerOpen && <div className="drawer-touch-blocker"/>
+         }
 
-        
-            
+       
+          
 
-          <MapContainer
-            center={defaultPosition}
-            zoom={defaultZoom}
-            style={{ height: "100%", width: "100%" }}
-            worldCopyJump={true}
-            zoomControl={false}
-            ref={setMap}
-            whenReady={() => {
-              console.log("Map is fully ready!");
-              
-            }}
+        <MapContainer
+          center={defaultPosition}
+          zoom={defaultZoom}
+          style={{ height: "100%", width: "100%" }}
+          worldCopyJump={true}
+          zoomControl={false}
+          ref={setMap}
+          whenReady={() => {
+            console.log("Map is fully ready!");
+            
+          }}
 
-            
-            
-          > <CustomZoomControl /> 
-            
-            
+          
+          
+        > <CustomZoomControl /> 
+          
+          
 
-            <TileLayer
-              url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-              attribution="&copy; <a href=&quot;http://osm.org/copyright&quot;>OpenStreetMap</a> contributors"
-            />
+          <TileLayer
+            url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+            attribution="&copy; <a href=&quot;http://osm.org/copyright&quot;>OpenStreetMap</a> contributors"
+          />
 
-            <LayersControl position="bottomright">
-              <LayersControl.Overlay checked name="ECCC Hurricane Response Zone">
-                <LayerGroup>
-                  <WMSTileLayer
-                    url="https://geo.weather.gc.ca/geomet"
-                    layers='HURRICANE_RESPONSE_ZONE'
-                    format='image/png'
-                    transparent='true'
-                    styles='HURRICANE_LINE_BLACK_DASHED'
-                    attribution='<a href=&quot;https://www.canada.ca/en/environment-climate-change.html&quot;>ECCC</a>'
-                    version='1.3.0'
-                  />
-                </LayerGroup>
-              </LayersControl.Overlay>
-              <LayersControl.Overlay checked name="Stations">
-                <LayerGroup>
-                  {
-                    station_data ? (
-                      Object.entries(station_data).map((station) => {
-                        let storm_timestamp = new Date()
-                        if("TIMESTAMP" in state.hover_marker.properties){
-                          storm_timestamp = new Date(["TIMESTAMP"]);
-                        }
-                        console.log(JSON.stringify(state.hover_marker))
-                        console.log(storm_timestamp)
-                        console.log(station)
-                        return (
-                          <StationMarker
-                            key={station[0]}
-                            station_data={station}
-                            time={storm_timestamp}
-                            selected_station={state.selected_station}
-                            dispatch={dispatch} 
-                      />)
-                        
-                      })
-                    ) : (
-                      <></>
-                    )
-                  }
-                </LayerGroup>
-              </LayersControl.Overlay>
-              <LayersControl.Overlay checked name="Error Cone">
-                <LayerGroup>
-                  {
-                    state.storm_points?.err?.features?.map(err_cone => {
+          <LayersControl position="bottomright">
+            <LayersControl.Overlay checked name="ECCC Hurricane Response Zone">
+              <LayerGroup>
+                <WMSTileLayer
+                  url="https://geo.weather.gc.ca/geomet"
+                  layers='HURRICANE_RESPONSE_ZONE'
+                  format='image/png'
+                  transparent='true'
+                  styles='HURRICANE_LINE_BLACK_DASHED'
+                  attribution='<a href=&quot;https://www.canada.ca/en/environment-climate-change.html&quot;>ECCC</a>'
+                  version='1.3.0'
+                />
+              </LayerGroup>
+            </LayersControl.Overlay>
+            <LayersControl.Overlay checked name="Stations">
+              <LayerGroup>
+                {
+                  station_data ? (
+                    Object.entries(station_data).map((station) => {
+                      let storm_timestamp = new Date()
+                      if("TIMESTAMP" in state.hover_marker.properties){
+                        storm_timestamp = new Date(["TIMESTAMP"]);
+                      }
+                      console.log(JSON.stringify(state.hover_marker))
+                      console.log(storm_timestamp)
+                      console.log(station)
                       return (
-                        <ErrorCone
-                          key={err_cone.id}
-                          error_cone_data={err_cone}
-                        />
-                      );
-                    })
-                  }
-                </LayerGroup>
-              </LayersControl.Overlay>
-              <LayersControl.Overlay checked name="Points">
-                <LayerGroup>
-                  {
-                    state.storm_points?.pts?.features?.map(point => {
-                      return (
-
-                        <StormMarker
-                          key={point.id}
-                          storm_point_data={point}
-                          storm_point_hover= {state.hover_marker}
-                          dispatch={dispatch}
-
-                        />
-                      );
-                    })
-                  }
-                </LayerGroup>
-              </LayersControl.Overlay>
-              <LayersControl.Overlay checked name="Track Line">
-                <LayerGroup>
-                  { 
-                    state.storm_points?.lin?.features?.length > 0 &&
-                    state.storm_points?.lin?.features?.map(line => {
+                        <StationMarker
+                          key={station[0]}
+                          station_data={station}
+                          time={storm_timestamp}
+                          selected_station={state.selected_station}
+                          dispatch={dispatch} 
+                    />)
                       
-
-                      return (
-                        <LineOfTravel
-                          key={line.id}
-                          storm_line_data={line}
-                        />
-                      );
                     })
-                  }
-                </LayerGroup>
-              </LayersControl.Overlay>
-              <LayersControl.Overlay checked name="Wind Speed Radius">
-                <LayerGroup>
-                  {
-                    state.storm_points?.rad?.features?.length > 0 &&
-                    state.storm_points?.rad?.features?.map(radii => {
-                      return (
-                        <WindSpeedRadius
-                          key={radii.id}
-                          storm_wind_radii_data={radii}
-                          hover_marker={state.hover_marker}
-                        />
-                      );
-                    })
-                  }
-                </LayerGroup>
-              </LayersControl.Overlay>
-              <LayersControl.Overlay checked name="Sea Height Radius">
-                <LayerGroup>
-                  {
-                    state.storm_points?.sea?.features?.length > 0 &&
-                    state.storm_points?.sea?.features?.map(radii => {
-                      return (
-                        <SeaHeightRadius
-                          key={radii.id}
-                          storm_sea_height_data={radii}
-                          hover_marker={state.hover_marker}
-                        />
-                      );
-                    })
-                  }
-                </LayerGroup>
-              </LayersControl.Overlay>
-            </LayersControl>
+                  ) : (
+                    <></>
+                  )
+                }
+              </LayerGroup>
+            </LayersControl.Overlay>
+            <LayersControl.Overlay checked name="Error Cone">
+              <LayerGroup>
+                {
+                  state.storm_points?.err?.features?.map(err_cone => {
+                    return (
+                      <ErrorCone
+                        key={err_cone.id}
+                        error_cone_data={err_cone}
+                      />
+                    );
+                  })
+                }
+              </LayerGroup>
+            </LayersControl.Overlay>
+            <LayersControl.Overlay checked name="Points">
+              <LayerGroup>
+                {
+                  state.storm_points?.pts?.features?.map(point => {
+                    return (
 
-            {<RenderSpatialFilter
-            ref={clearShapesRef} 
-            setPolyFilterCoords={(coords) => dispatch({ type: "SET_POLY_FILTER_COORDS", payload: coords })}
-            />} {/* Calling the EditControl function here */}
-          </MapContainer>
+                      <StormMarker
+                        key={point.id}
+                        storm_point_data={point}
+                        storm_point_hover= {state.hover_marker}
+                        dispatch={dispatch}
 
-          { map && (<Drawer
-              element_id="left-side"
-              classes="left"
-              source_type={source_type}
-              setStationPoints={setStationPoints}
-              state={state}
-              dispatch={dispatch}
-              map={map}
-            />)}
-        </div>
+                      />
+                    );
+                  })
+                }
+              </LayerGroup>
+            </LayersControl.Overlay>
+            <LayersControl.Overlay checked name="Track Line">
+              <LayerGroup>
+                { 
+                  state.storm_points?.lin?.features?.length > 0 &&
+                  state.storm_points?.lin?.features?.map(line => {
+                    
+
+                    return (
+                      <LineOfTravel
+                        key={line.id}
+                        storm_line_data={line}
+                      />
+                    );
+                  })
+                }
+              </LayerGroup>
+            </LayersControl.Overlay>
+            <LayersControl.Overlay checked name="Wind Speed Radius">
+              <LayerGroup>
+                {
+                  state.storm_points?.rad?.features?.length > 0 &&
+                  state.storm_points?.rad?.features?.map(radii => {
+                    return (
+                      <WindSpeedRadius
+                        key={radii.id}
+                        storm_wind_radii_data={radii}
+                        hover_marker={state.hover_marker}
+                      />
+                    );
+                  })
+                }
+              </LayerGroup>
+            </LayersControl.Overlay>
+            <LayersControl.Overlay checked name="Sea Height Radius">
+              <LayerGroup>
+                {
+                  state.storm_points?.sea?.features?.length > 0 &&
+                  state.storm_points?.sea?.features?.map(radii => {
+                    return (
+                      <SeaHeightRadius
+                        key={radii.id}
+                        storm_sea_height_data={radii}
+                        hover_marker={state.hover_marker}
+                      />
+                    );
+                  })
+                }
+              </LayerGroup>
+            </LayersControl.Overlay>
+          </LayersControl>
+
+          {<RenderSpatialFilter
+          ref={clearShapesRef} 
+          setPolyFilterCoords={(coords) => dispatch({ type: "SET_POLY_FILTER_COORDS", payload: coords })}
+          />} {/* Calling the EditControl function here */}
+        </MapContainer>
+
+        { map && (<Drawer
+            element_id="left-side"
+            classes="left"
+            source_type={source_type}
+            setStationPoints={setStationPoints}
+            state={state}
+            dispatch={dispatch}
+            map={map}
+          />)}
       </div>
-    </TourProvider>
+    </div>
     
   )
 }
