+++ react/components/mapReducer.js
@@ -1,5 +1,4 @@
 // mapReducer.js
-import { is } from "date-fns/locale";
 import { empty_point_obj, empty_station_obj, empty_storm_obj } from "./point_defaults";
 
 
@@ -26,8 +25,7 @@ export const initialMapState = {
   filterStormName:[],
   selectedStormNames:[],
   info:true,
-  showFilterSelected:true,
-  isTour: true
+  showFilterSelected:true
 };
 
 export function mapReducer(state, action) {
@@ -72,8 +70,6 @@ export function mapReducer(state, action) {
       return { ...state, showCatSelection: !state.showCatSelection };
     case 'SET_CAT_SELECTION':
       return { ...state, showCatSelection: action.payload };
-    case 'SET_TOUR':
-      return { ...state, isTour: action.payload };
     case 'SET_DATE_SELECTION':
       return { ...state, showDateSelection: action.payload };
       case "TOGGLE_DATE_SELECTION":
